Refactor all game logic into game loop
Get game logic closer to how it will be for other graphical systems
and move control of the game data out of window classes by running
a game loop. We now separately update game logic and render the state
of the game world. Groups of windows are pushed onto a vector stack
and popped off as windows close.